Fixed the HDR playback black screen issue
The EDID parsed from Dell U3219Q with ADL NUC HDMI ports is not
correct and this results in the HDR video playback as black
screen. The fix is to parsing the HDR capability at connector
initialization stage and reports the HDR capabilities to SF
corectly in later HWC API calls.

Signed-off-by: Wan Shuang <shuang.wan@intel.com>
Tracked-On: OAM-106847