drm_hwcomposer: Treat unknown connector as internal display
The code determines if a display is internal or external based on the
display connector type, and at least one internal display must exist
which is used as the primary display. Because the FVP Base RevC does
not model any connector hardware, the reported connector type is none/
unknown, so we need to treat unknown as an internal connector.

Change-Id: Ifa603f903d74ccad6c7e9da13042994a5fdffe32
Signed-off-by: Lingkai Dong <lingkai.dong@arm.com>