drm_hwcomposer: Add pl111 to the client backend device list
PL111[1] is the color LCD controller available on the Armv-A Base
Platform RevC FVP and the Versatile Express LogicTile FPGA board. The
8MB video memory of PL111 is barely enough for the two fullscreen
buffers always requested by SurfaceFlinger for client composition, and
there is no video memory left for any device composition buffers.
Therefore we add PL111 to the list of devices that uses the client-only
composition backend.

[1]: https://developer.arm.com/Processors/PL111

Change-Id: Ib64ccf5e030b623d5c920eb998c6b22ead89b4c9
Signed-off-by: Lingkai Dong <lingkai.dong@arm.com>