drm_hwcomposer: Fix bug where DrmDevice can be released early
DrmFbImporter can outlive DrmDevice which will cause issues when
resources are released. Addressing this would require a restructure
on how ResourceManager stores DrmDevices and DrmFbImporter to make
sure they depend on each other. For now, just acquire the DRM fd from
the DrmDevice to make sure it is not closed.

Change-Id: If46ca0f879bee4a5da2f93d341b6f1278e593010
Signed-off-by: Normunds Rieksts <normunds.rieksts@arm.com>
Signed-off-by: Drew Davenport <ddavenport@google.com>