drm_hwcomposer: CI: publish docker image to local container registry
drm_hwcomposer provides a .ci/Dockerfile which can be used as a base
image to rebuild drm_hwcomposer.
Unfortunetly, the CI stages (build, tidy, style) do not use this
container image. Instead, it re-installs all the dependencies for
each stage.

To use this container image, we should also build it via the CI.

Add a new build stage to build the container and publish it to the
local (gitlab-provided) docker registry.

This way, it can be reused by the other build stages.

Note: as documented in [1], this requires that the configured Gitlab
runner can run docker in privileged mode.

[1] https://docs.gitlab.com/ee/ci/docker/using_docker_build.html#use-docker-in-docker
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>