drm_hwcomposer: work around for bad edid display size and dpi.
drm-hwc computes the dpi by dividing the resolution by the size of
the panel. That should technically be correct at full resolution,
however the panel size comes from the edid and in practice display
manufacturers are not the most reliable when populating edid. This is
leads to some wildly inaccurate dpi value that can change the aspect
ratio of the screen from 16:9 to 4:3.
Since edid are unreliable one simple work around is to only use one
dpi value, the width is the most likely to be correct as it is the less
likely to have bezels.

Change-Id: I8fce08dc137129b5176050bb2c61055571f1cca8