drm_hwcomposer: CI: Use upstream container image conditionally
Right now we use $CI_REGISTRY_IMAGE:latest as the docker image for
running build, tidy and checkstyle.

The problem with this is that CI_REGISTRY_IMAGE points to the *fork*'s
container registry.

For example, if the ddavenport user has this fork:
https://gitlab.freedesktop.org/ddavenport/drm-hwcomposer

Then CI_REGISTRY_IMAGE will be:
registry.freedesktop.org/ddavenport/drm-hwcomposer

This won't work for forks that never build the container image.

Use registry.freedesktop.org/drm-hwcomposer/drm-hwcomposer:latest as
container image by default.

For MRs that modify .ci/Dockerfile, a container will be rebuild
and the forks image will be used instead.

Link: https://gitlab.freedesktop.org/drm-hwcomposer/drm-hwcomposer/-/issues/97
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>