Revert "drm_hwcomposer: Workaround for screen de-activating causing db845c regression"
This workaround actually leaves pipes in an unclean state on external
displays which leads to planes bleeding and incorrect renderings in the
case of overlays being assigned to external displays.

There is also a case to be made to not have board specific hacks in
drm_hwc and we should fine a better solution if this is still a problem.
We could consider only deactivating non-primary display if the new
pipeline is set to null for instance.

This reverts commit 799e8c74c65e5d12d354a0b3f2b541f3d44273a2.

Change-Id: Ia28f9e3b0a9ee81ff934143bd2aba526219b76e4
Signed-off-by: Lucas Berthou <berlu@google.com>