drm_hwcomposer: Query property when checking virtual displays availability
The current implementation of "getMaxVirtualDisplayCount" function
relies on the number of writeback connectors, but doesn't consider
the availability of CRTCs. Since CRTCs can only be bound to a single
pipeline, even though there are available writeback connectors, we
might fail to create a virtual pipeline if the CRTCs are being used.

This leads to failures in the "CreateVirtualDisplay" and
"SetOutputBuffer" VTS tests because the tests expect the creation of
a virtual pipeline to work based on the positive value returned by
"getMaxVirtualDisplayCount".

The virtual display feature in drm-hwcomposer is experimental and
currently unused. To address test failures, modify the function
"getMaxVirtualDisplayCount" to return 0 unless the property
"vendor.vendor.hwc.drm.enable_virtual_display" is explicitly set to
true.

Signed-off-by: Paz Zcharya <pazz@google.com>