drm_hwcomposer: Fix DrmConnector log build failure
This fix is the same as the previous commit on DrmProperty:
    2c36651af0ad ("drm_hwcomposer: Fix DrmProperty log build failure")

So copy the commit message of it here for reference convenience.

DrmConnector logs the uint64_t property value if it is unrecognized.
However, this type can be represented multiple ways:

* unsigned long (%lu)
* unsigned long long (%llu)

This generates build errors when the wrong type is used.

The correct way to log the value is with |PRIu64|.

Change-Id: Ie59ac6930919f77c16c064f3d5becfa794ed37a3
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>