drm_hwcomposer: Add blocking SetConfig
Config changes within the same config group are intended to change
the refresh rate and are expected to be seamless, and as such are queued
up to be applied along with other frame updates such as pageflips. These
commits ought not to result in a full modeset.

For config changes that go between config groups, there might be visible
artifacts (modesets).

Introduce HwcDisplay::SetConfig to implement support for changing
between config groups. Config changes that are expected to be seamless
and not introduce jank or other visual artifacts should continue to go
through HwcDisplay::QueueConfig.

Since there may not be an appropriate client buffer available for the
initial SetConfig, allocate a CPU-writable buffer for scanout that can
be used for the first commit. The buffer can be destroyed after the
next frame is presented.

Change-Id: I48a87a070130dd2328f415719032486d9659c5b6
Signed-off-by: Drew Davenport <ddavenport@google.com>