drm_hwcomposer: Check "Present Not Reliable" property
Check the "Present Not Reliable" property to determine if the capability
PRESENT_FENCE_IS_NOT_RELIABLE should be enabled or not.

The virtio gpu module sends frames to the host as fast as possible and
does not emulate "real display timing". Devices using the virtio gpu
module (for example, the virtual Android device Cuttlefish) should set
the following property to "true" to enable the capability:

    ro.vendor.hwc.drm.present_not_reliable.enabled

For example, in an Android BoardConfig.mk:

    PRODUCT_VENDOR_PROPERTIES += \
        vendor.hwc.drm.present_not_reliable.enabled=true

Signed-off-by: Tim Van Patten <timvp@google.com>