drm_hwcomposer: Define cc_defaults for hwc3 binary
Some vendors do not use gralloc4 and can't use the BufferInfoGetter
implemented on top of the IMapper 4 API. These vendors use a
LegacyBufferInfoGetter implementation, and need to define a new build
target for it.

Move the majority of the configuration for the hwc3 service to a
cc_defaults target, so that this can be referenced downstream and reduce
duplication.

Change-Id: I2561ecbb2d3c520f8cd8bdbf171a28d5f7fb1a51
Signed-off-by: Drew Davenport <ddavenport@google.com>