drm_hwcomposer: HWC3: Fix hotplug handling
The second or any subsequent hotplug event with connected status enabled is
a special way to inform the client (SF) that the display has changed its
dimensions. In this case, the client removes all layers and re-creates
them. In this case, we keep the display resources.

This avoids logcat messages during plug/unplug the primary display:

    E HwcComposer: destroyLayer failed Status(-8, EX_SERVICE_SPECIFIC): '3: '
    E HWC2    : destroyLayer(0, 3) failed: BadLayer (3)

Change-Id: I24b110a0b474cd56de0d51b83b0fb44d79d3fccd
Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>