drm_hwcomposer: Set vsync callback on demand
Set and clear a vsync callback based on when the client requests vsync
events to be enabled or not. The new vsync callback does not need to
acquire the global resource lock.

The vsync callback uses the vsync period that was set in the VSyncWorker
rather than querying it from the HwcDisplay.

Change-Id: Ic7868af4963ad40aa5a0982a190c67f1642d5b32
Signed-off-by: Drew Davenport <ddavenport@google.com>