drm_hwcomposer: Simplify default vsync period
VSyncWorker::SetVsyncPeriod won't be called with a value of 0, so the
only time the internal vsync period will be zero is if SetVsyncPeriod
was not called.

Set the 60Hz vsync period as a default value, and let the synthetic
vsync code expect it to be set to a reasonable value.

Change-Id: Ie1f83b0c11504acf304cc241923bfa60a5ffd647
Signed-off-by: Drew Davenport <ddavenport@google.com>