Support virtio-GPU backed by both iGPU and dGPU scanouts (v2)
Add a flag to the HwcLayer which will be set when the layer is allowed to be
shared as local memory objects without migrating to system memory, primarily
used in the case that the layer is to be displayed on dGPU. If this flag is
enabled, we will set the DMA BUF name to "p2p" via DMA_BUF_SET_NAME ioctl. The
name acts as a hint in the KMD for local memory sharing.

This feature cannot be achieved without the help from virtio-GPU KMD: now
virtio-GPU will expose a bit mask (queried by ALLOW_P2P parameter) indicating
whether local memory sharing is allowed for each scan-out.

Change log:
- v2: combine commit "Avoid dereferencing null pointer in headless mode"

Tracked-On: OAM-129252
Signed-off-by: Weifeng Liu <weifeng.liu@intel.com>