drm_hwcomposer: CI: Make checkstyle a placeholder
The checkstyle job runs a quick script but currently makes use of a full
gitlab runner. Instead, add the placeholder tag to this job so that it
can be picked up and run in parallel with other light-weight jobs. This
should free up resources and avoid waits to run the checkstyle job.

Signed-off-by: Deborah Brouwer <deborah.brouwer@collabora.com>