drm_hwcomposer: Do not use config groups by default
When config groups are enabled, the framework uses the Multiple Refresh
Rate feature, which causes blank screen issues on some devices, making
them unusable.

This is a temporary solution until we find a reliable way to query the
driver if a seamless mode change is supported.

Discussion can be found [here][1]

[1]: https://gitlab.freedesktop.org/drm-hwcomposer/drm-hwcomposer/-/merge_requests/295#note_2770300
Change-Id: If17c29de9f2cd853887ece5fe050e4ceee5ddb65
Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>