drm_hwcomposer: Fix deadlock with flattening
FlatteningController acquires its internal lock first, and then acquires
the global lock in DrmHwcThree::SendRefreshEventToClient.

During normal validate/present, the thread acquires the global lock
first, and will acquire the FlatteningController internal lock in
NewFrame.

Move the "must_validate" flag out of Hwc3Display to remove the need to
acquire the global lock in the SendRefreshEventToClient callback.
Add a std::set to track which displays must be validated, and protect
access to this set with a different mutex to avoid deadlocks.

Tracked-On: OAM-133858
Change-Id: I62d59252107f9f16a4b8bb43f850115bf327a127
Signed-off-by: Drew Davenport <ddavenport@google.com>
Signed-off-by: ZhuChenyanX <zhucx@intel.com>